refactor: promote sqlite to top-level submodule
Move src/storage/sqlite/ to src/sqlite-storage/ and rename
storage-sqlite.ts to main.ts so it follows the same shape as the other
submodules. Update package.json export and all importers accordingly.

This lets check-imports.ts drop its package-exports exception: every
cross-submodule import now uniformly targets <module>/main.ts. Also
move check-imports invocation from `npm run check` (which is for
type-checking) to `npm run lint` (which is for static rules).
