+++ scripts/check-imports.ts
@@ -4,12 +4,6 @@ import { dirname, relative, resolve } from "node:path";
 const ROOT = resolve(import.meta.dirname, "..");
 const SRC = resolve(ROOT, "src");
 
-const pkg = JSON.parse(await readFile(resolve(ROOT, "package.json"), "utf8"));
-const publicEntries = new Set<string>();
-for (const value of Object.values(pkg.exports as Record<string, string>)) {
-  publicEntries.add(resolve(ROOT, value.replace(/^\.\/dist\//, "src/").replace(/\.js$/, ".ts")));
-}
-
 const importRegex = /(?:from|import)\s+["']([^"']+)["']/g;
 const violations: string[] = [];
 
@@ -31,12 +25,8 @@ for await (const file of glob("src/**/*.ts", { cwd: ROOT })) {
     if (targetModule === moduleOfFile) {
       continue;
     }
-    const isMain = relFromSrc === `${targetModule}/main.ts`;
-    const isPublic = publicEntries.has(targetAbs);
-    if (!isMain && !isPublic) {
-      violations.push(
-        `${file}: imports "${spec}" — cross-submodule imports must target ../${targetModule}/main.ts (or a package export)`,
-      );
+    if (relFromSrc !== `${targetModule}/main.ts`) {
+      violations.push(`${file}: imports "${spec}" — cross-submodule imports must target ../${targetModule}/main.ts`);
     }
   }
 }
