+++ scripts/serve-mailbox.ts
@@ -5,7 +5,7 @@ import { ed25519 } from "@noble/curves/ed25519.js";
 import { scrypt } from "@noble/hashes/scrypt.js";
 import { createMailboxRouter, Mailbox } from "../src/mailbox/main.ts";
 import { createListener } from "../src/nodejs-utils/serve.ts";
-import { NodeSqliteDatabase, SqliteControllerStorage } from "../src/storage/sqlite/storage-sqlite.ts";
+import { NodeSqliteDatabase, SqliteControllerStorage } from "../src/sqlite-storage/main.ts";
 
 const storage = new SqliteControllerStorage(new NodeSqliteDatabase(new DatabaseSync(":memory:")));
