+++ src/sqlite-storage/main.test.ts
@@ -2,8 +2,8 @@ import assert from "node:assert/strict";
 import { basename } from "node:path";
 import { DatabaseSync } from "node:sqlite";
 import { describe, test } from "node:test";
-import { type InceptEventBody, type KeyEvent, type KeyState, keri, Message } from "../../core/main.ts";
-import { NodeSqliteDatabase, SqliteControllerStorage } from "./storage-sqlite.ts";
+import { type InceptEventBody, type KeyEvent, type KeyState, keri, Message } from "../core/main.ts";
+import { NodeSqliteDatabase, SqliteControllerStorage } from "./main.ts";
 
 function incept(): KeyEvent<InceptEventBody> {
   const key0 = keri.utils.generateKeyPair();
