+++ src/witness/witness.test.ts
@@ -5,7 +5,7 @@ import { describe, test } from "node:test";
 import { ed25519 } from "@noble/curves/ed25519.js";
 import { encodeText, Indexer, Matter, Message } from "../cesr/main.ts";
 import { generateKeyPair, KeyEventLog, keri, verifySignature } from "../core/main.ts";
-import { NodeSqliteDatabase, SqliteControllerStorage } from "../storage/sqlite/storage-sqlite.ts";
+import { NodeSqliteDatabase, SqliteControllerStorage } from "../sqlite-storage/main.ts";
 import { Witness, WitnessError } from "./witness.ts";
 
 function makeWitness(seed = "test-witness") {
